implemented broadcast settings to notification manager
